Better way to share data between ncurses and sdl2.
