+++ source/frontends/common2/config.h.in
