+++ source/frontends/common2/configuration.h
