+++ source/frontends/common2/resources.h
