+++ source/frontends/common2/utils.h
@@ -1,0 +1,6 @@
+#include <string>
+#include <linux/windows/wincompat.h>
+
+bool DoDiskInsert(const UINT slot, const int nDrive, const std::string & fileName, const bool createMissingDisk);
+
+void setSnapshotFilename(const std::string & filename);
