+++ source/frontends/ncurses/CMakeLists.txt
@@ -1,5 +1,4 @@
 include(FindPkgConfig)
-include(GNUInstallDirs)
 
 add_executable(applen
   main.cpp
@@ -8,23 +7,16 @@ add_executable(applen
   evdevpaddle.cpp
   nframe.cpp
   asciiart.cpp
-  resources.cpp
-  configuration.cpp
   bitmaps.cpp
   )
 
 pkg_search_module(NCURSESW REQUIRED ncursesw)
 pkg_search_module(LIBEVDEV REQUIRED libevdev)
-find_package(Boost REQUIRED
-  COMPONENTS program_options
-  )
 
 target_include_directories(applen PRIVATE
-  ${CMAKE_CURRENT_BINARY_DIR}
   ${NCURSESW_INCLUDE_DIRS}
   ${LIBEVDEV_INCLUDE_DIRS}
   ${Boost_INCLUDE_DIRS}
-  ${CMAKE_CURRENT_BINARY_DIR}
   )
 
 target_compile_options(applen PRIVATE
@@ -35,17 +27,9 @@ target_compile_options(applen PRIVATE
 target_link_libraries(applen PRIVATE
   ${NCURSESW_LIBRARIES}
   ${LIBEVDEV_LIBRARIES}
-  Boost::program_options
   appleii
+  common2
   )
-
-file(RELATIVE_PATH ROOT_PATH ${CMAKE_BINARY_DIR} ${CMAKE_SOURCE_DIR})
-file(RELATIVE_PATH SHARE_PATH ${CMAKE_INSTALL_FULL_BINDIR} ${CMAKE_INSTALL_FULL_DATADIR}/applewin)
-
-configure_file(config.h.in config.h)
 
 install(TARGETS applen
   DESTINATION bin)
-
-install(DIRECTORY ${CMAKE_SOURCE_DIR}/resource
-  DESTINATION ${CMAKE_INSTALL_DATADIR}/applewin)
