+++ source/frontends/sa2/CMakeLists.txt
@@ -3,8 +3,6 @@ include(FindPkgConfig)
 add_executable(sa2
   main.cpp
   bitmaps.cpp
-  ../ncurses/configuration.cpp
-  ../ncurses/resources.cpp
   )
 
 find_package(Boost REQUIRED
@@ -18,13 +16,13 @@ target_compile_features(sa2 PUBLIC cxx_std_17)
 target_include_directories(sa2 PRIVATE
   ${Boost_INCLUDE_DIRS}
   ${SDL2_INCLUDE_DIRS}
-  ${CMAKE_CURRENT_BINARY_DIR}
   )
 
 target_link_libraries(sa2 PRIVATE
   Boost::program_options
   ${SDL2_LIBRARIES}
   appleii
+  common2
   )
 
 install(TARGETS sa2
